INGM-638 created frame.py, calling validator from PDUMaps
