+++ examples/safety_mapping_example.py
@@ -14,10 +14,6 @@ if FSOE_MASTER_INSTALLED:
         STOFunction,
         SVFunction,
     )
-    from ingeniamotion.fsoe_master.maps_validator import (
-        FSoEDictionaryMapValidator,
-        FSoEFrameConstructionError,
-    )
 
 
 def main(interface_ip, slave_id, dict_path) -> None:
@@ -71,17 +67,7 @@ def main(interface_ip, slave_id, dict_path) -> None:
     inputs.add_padding(7)
 
     # Check that the maps are valid
-    validator = FSoEDictionaryMapValidator()
-    for dictionary_map, map_description in zip(
-        [handler.maps.inputs, handler.maps.outputs], ["Inputs Map", "Outputs Map"]
-    ):
-        validator.reset()
-        print(f"Validating {map_description}")
-        exceptions = validator.validate(dictionary_map, rules_to_validate=None)
-        if exceptions:
-            print(f"Validation failed for {map_description}:")
-            raise FSoEFrameConstructionError(exceptions)
-        print(f"{map_description} is valid.")
+    handler.maps.is_map_valid(raise_exceptions=True)
 
     # Print the maps to check the configuration
     print("Inputs Map:")
